add hook for wallet filtering
