+++ packages/modal/src/ui/containers/ConnectWallet/ConnectWallet.tsx
@@ -1,11 +1,12 @@
-import { ANALYTICS_EVENTS, type ChainNamespaceType, log, type WALLET_CONNECTOR_TYPE, WALLET_CONNECTORS } from "@web3auth/no-modal";
-import { FormEvent, useCallback, useContext, useEffect, useMemo, useState } from "react";
+import { ANALYTICS_EVENTS, log, WALLET_CONNECTORS } from "@web3auth/no-modal";
+import { useCallback, useContext, useEffect, useMemo, useState } from "react";
 
 import { CONNECT_WALLET_PAGES, PAGES } from "../../constants";
 import { AnalyticsContext } from "../../context/AnalyticsContext";
 import { useModalState } from "../../context/ModalStateContext";
 import { useBodyState } from "../../context/RootContext";
 import { useWidget } from "../../context/WidgetContext";
+import { useWalletFiltering } from "../../hooks/useWalletFiltering";
 import { ExternalButton } from "../../interfaces";
 import { ConnectWalletProps } from "./ConnectWallet.type";
 import ConnectWalletChainFilter from "./ConnectWalletChainFilter";
@@ -22,15 +23,31 @@ function ConnectWallet(props: ConnectWalletProps) {
   const { deviceDetails, isDark, uiConfig } = useWidget();
   const { walletRegistry } = uiConfig;
   const { modalState, setModalState, handleShowExternalWallets, preHandleExternalWalletClick: handleExternalWalletClick } = useModalState();
-  const { externalWalletsConfig: config, walletConnectUri, metamaskConnectUri } = modalState;
+  const { walletConnectUri, metamaskConnectUri } = modalState;
+
+  const {
+    allUniqueButtons,
+    externalButtons,
+    totalExternalWalletsCount,
+    initialWalletCount,
+    walletSearch,
+    handleWalletSearch,
+    selectedChain,
+    setSelectedChain: handleChainFilterChange,
+    isShowAllWallets,
+    handleMoreWallets,
+  } = useWalletFiltering({
+    allRegistryButtons,
+    customConnectorButtons,
+    connectorVisibilityMap,
+    externalWalletsConfig: modalState.externalWalletsConfig,
+    walletRegistry,
+  });
 
   const [currentPage, setCurrentPage] = useState(CONNECT_WALLET_PAGES.CONNECT_WALLET);
   const [selectedWallet, setSelectedWallet] = useState(false);
   const [isLoading] = useState<boolean>(false);
   const [selectedButton, setSelectedButton] = useState<ExternalButton>(null);
-  const [walletSearch, setWalletSearch] = useState<string>("");
-  const [selectedChain, setSelectedChain] = useState<string>("all");
-  const [isShowAllWallets, setIsShowAllWallets] = useState<boolean>(false);
   // Track if user came directly from Login page with a pre-selected wallet
   const [isPreSelectedFromLogin, setIsPreSelectedFromLogin] = useState(false);
 
@@ -68,114 +85,6 @@ function ConnectWallet(props: ConnectWalletProps) {
   const disableBackButton = useMemo(() => {
     return bodyState.installLinks?.show || bodyState.multiChainSelector?.show;
   }, [bodyState.installLinks?.show, bodyState.multiChainSelector?.show]);
-
-  const walletDiscoverySupported = useMemo(() => {
-    const supported = walletRegistry && (Object.keys(walletRegistry.default || {}).length > 0 || Object.keys(walletRegistry.others || {}).length > 0);
-    return supported;
-  }, [walletRegistry]);
-
-  const allUniqueButtons = useMemo(() => {
-    const uniqueButtonSet = new Set();
-    return customConnectorButtons.concat(allRegistryButtons).filter((button) => {
-      if (uniqueButtonSet.has(button.name)) return false;
-      uniqueButtonSet.add(button.name);
-      return true;
-    });
-  }, [allRegistryButtons, customConnectorButtons]);
-
-  const defaultButtonKeys = useMemo(() => new Set(Object.keys(walletRegistry.default)), [walletRegistry]);
-
-  const defaultButtons = useMemo(() => {
-    // display order: default injected buttons > custom adapter buttons > default non-injected buttons
-    const buttons = [
-      ...allRegistryButtons.filter((button) => button.hasInjectedWallet && defaultButtonKeys.has(button.name)),
-      ...customConnectorButtons,
-      ...allRegistryButtons.filter((button) => !button.hasInjectedWallet && defaultButtonKeys.has(button.name)),
-    ].sort((a, b) => {
-      // favor MetaMask over other wallets
-      if (a.name === WALLET_CONNECTORS.METAMASK && b.name === WALLET_CONNECTORS.METAMASK) {
-        // favor installed MetaMask over non-installed MetaMask
-        if (a.isInstalled) return -1;
-        if (b.isInstalled) return 1;
-        // favor injected MetaMask over non-injected MetaMask
-        if (a.hasInjectedWallet) return -1;
-        if (b.hasInjectedWallet) return 1;
-        return 0;
-      }
-      if (a.name === WALLET_CONNECTORS.METAMASK) return -1;
-      if (b.name === WALLET_CONNECTORS.METAMASK) return 1;
-      return 0;
-    });
-
-    const buttonSet = new Set();
-    return buttons
-      .filter((button) => {
-        if (buttonSet.has(button.name)) return false;
-        buttonSet.add(button.name);
-        return true;
-      })
-      .filter((button) => selectedChain === "all" || button.chainNamespaces?.includes(selectedChain as ChainNamespaceType));
-  }, [allRegistryButtons, customConnectorButtons, defaultButtonKeys, selectedChain]);
-
-  const installedWalletButtons = useMemo(() => {
-    const visibilityMap = connectorVisibilityMap;
-    return Object.keys(config).reduce((acc, localConnector) => {
-      if (localConnector !== WALLET_CONNECTORS.WALLET_CONNECT_V2 && visibilityMap[localConnector]) {
-        acc.push({
-          name: localConnector,
-          displayName: config[localConnector as WALLET_CONNECTOR_TYPE].label || localConnector,
-          hasInjectedWallet: config[localConnector as WALLET_CONNECTOR_TYPE].isInjected,
-          hasWalletConnect: false,
-          hasInstallLinks: false,
-        });
-      }
-      return acc;
-    }, [] as ExternalButton[]);
-  }, [connectorVisibilityMap, config]);
-
-  const handleWalletSearch = (e: FormEvent<HTMLInputElement>) => {
-    const searchValue = (e.target as HTMLInputElement).value;
-    setWalletSearch(searchValue);
-  };
-
-  const handleChainFilterChange = (chain: string) => {
-    setSelectedChain(chain);
-  };
-
-  const filteredButtons = useMemo(() => {
-    if (walletDiscoverySupported) {
-      return [...allUniqueButtons.filter((button) => button.hasInjectedWallet), ...allUniqueButtons.filter((button) => !button.hasInjectedWallet)]
-        .sort((a, _) => (a.name === WALLET_CONNECTORS.METAMASK ? -1 : 1))
-        .filter((button) => selectedChain === "all" || button.chainNamespaces.includes(selectedChain as ChainNamespaceType))
-        .filter((button) => button.name.toLowerCase().includes(walletSearch.toLowerCase()));
-    }
-    return installedWalletButtons;
-  }, [walletDiscoverySupported, installedWalletButtons, walletSearch, allUniqueButtons, selectedChain]);
-
-  const externalButtons = useMemo(() => {
-    if (walletDiscoverySupported && !walletSearch && !isShowAllWallets) {
-      return defaultButtons;
-    }
-    return filteredButtons;
-  }, [walletDiscoverySupported, walletSearch, filteredButtons, defaultButtons, isShowAllWallets]);
-
-  const totalExternalWalletsCount = useMemo(() => filteredButtons.length, [filteredButtons]);
-
-  const initialWalletCount = useMemo(() => {
-    if (isShowAllWallets) return totalExternalWalletsCount;
-    return walletDiscoverySupported ? defaultButtons.length : installedWalletButtons.length;
-  }, [walletDiscoverySupported, defaultButtons, installedWalletButtons, isShowAllWallets, totalExternalWalletsCount]);
-
-  // Automatically show all wallets if there are less than or equal to 15 wallets
-  // also resets everytime we search causing no. of wallets to change or select different chain
-  useEffect(() => {
-    if (walletDiscoverySupported && totalExternalWalletsCount <= 15) {
-      // eslint-disable-next-line react-hooks/set-state-in-effect -- intentional
-      setIsShowAllWallets(true);
-    } else {
-      setIsShowAllWallets(false);
-    }
-  }, [walletDiscoverySupported, selectedChain, totalExternalWalletsCount]);
 
   // Handle pre-selected wallet from Login page (e.g., MetaMask QR code flow)
   useEffect(() => {
@@ -260,10 +169,6 @@ function ConnectWallet(props: ConnectWalletProps) {
         });
       }
     }
-  };
-
-  const handleMoreWallets = () => {
-    setIsShowAllWallets(true);
   };
 
   const qrCodeValue = useMemo(() => {
